update: rename organisms to sections
