+++ src/pages/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import PresentationSection from '../organisms/PresentationSection';
+import PresentationSection from '../sections/PresentationSection';
 
 function Home() {
   return (
